fixes in validations to pass Postman tests

+++ main-service/src/main/java/ru/practicum/service/category/CategoryServiceImpl.java
@@ -6,10 +6,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.exception.InvalidRequestException;
 import ru.practicum.exception.NotFoundException;
 import ru.practicum.mapper.CategoryMapper;
 import ru.practicum.model.Category;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,23 +20,31 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
+        validateAvailableName(newCategoryDto.getName());
         Category created = categoryRepository.save(CategoryMapper.mapToCategory(newCategoryDto));
         return CategoryMapper.mapToDto(created);
     }
 
     @Override
     public void deleteCategory(Long catId) {
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new InvalidRequestException("Нельзя удалить связанную с событием категорию");
+        }
         getCategoryOrThrow(catId);
-        //Сделать проверку, не связано ли с категориями событие
         categoryRepository.deleteById(catId);
     }
 
     @Override
     public CategoryDto updateCategory(NewCategoryDto newCategoryDto, Long catId) {
         Category category = getCategoryOrThrow(catId);
+        if(category.getName().equals(newCategoryDto.getName())) {
+            return CategoryMapper.mapToDto(category);
+        }
+        validateAvailableName(newCategoryDto.getName());
         category.setName(newCategoryDto.getName());
         Category saved = categoryRepository.save(category);
         return CategoryMapper.mapToDto(saved);
@@ -57,5 +67,11 @@ public class CategoryServiceImpl implements CategoryService {
     private Category getCategoryOrThrow(Long id) {
         return categoryRepository.findById(id).orElseThrow(() ->
                 new NotFoundException(String.format("Категория с id %d не существует", id)));
+    }
+
+    private void validateAvailableName(String name) {
+        if (categoryRepository.existsByName(name)) {
+            throw new InvalidRequestException(String.format("Имя категории %s уже занято", name));
+        }
     }
 }
